# db: change LazyFetcher.ValueFetcher to an interface