--- a/internal/base/lazy_value.go
+++ b/internal/base/lazy_value.go
@@ -152,8 +152,20 @@ type LazyValue struct {
 
 // LazyFetcher supports fetching a lazy value.
 type LazyFetcher struct {
-	// ValueFetcher, given a handle, returns the value. It is acceptable to call
-	// the ValueFetcher as long as the DB is open. However, one should assume
+	// Fetcher, given a handle, returns the value.
+	Fetcher ValueFetcher
+	// Attribute includes the short attribute and value length.
+	Attribute   AttributeAndLen
+	fetched     bool
+	value       []byte
+	callerOwned bool
+	err         error
+}
+
+// ValueFetcher is an interface for fetching a value.
+type ValueFetcher interface {
+	// Fetch returns the value, given the handle. It is acceptable to call the
+	// ValueFetcher.Fetch as long as the DB is open. However, one should assume
 	// there is a fast-path when the iterator tree has not moved off the sstable
 	// iterator that initially provided this LazyValue. Hence, to utilize this
 	// fast-path the caller should try to decide whether it needs the value or
@@ -163,14 +175,8 @@ type LazyFetcher struct {
 	// If the fetcher attempted to use buf *and* len(buf) was insufficient, it
 	// will allocate a new slice for the value. In either case it will set
 	// callerOwned to true.
-	ValueFetcher func(
+	Fetch(
 		handle []byte, valLen int32, buf []byte) (val []byte, callerOwned bool, err error)
-	// Attribute includes the short attribute and value length.
-	Attribute   AttributeAndLen
-	fetched     bool
-	value       []byte
-	callerOwned bool
-	err         error
 }
 
 // Value returns the underlying value.
@@ -195,7 +201,7 @@ func (lv *LazyValue) fetchValue(buf []byte) (val []byte, callerOwned bool, err e
 	f := lv.Fetcher
 	if !f.fetched {
 		f.fetched = true
-		f.value, f.callerOwned, f.err = lv.Fetcher.ValueFetcher(
+		f.value, f.callerOwned, f.err = f.Fetcher.Fetch(
 			lv.ValueOrHandle, lv.Fetcher.Attribute.ValueLen, buf)
 	}
 	return f.value, f.callerOwned, f.err